crawler/requirements.txt
beautifulsoup4==4.7.1
cffi==1.11.5
cld2-cffi==0.1.4
decorator==4.0.11
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
multi-rake==0.0.1
numpy==1.16.1
Pillow==6.0.0
pycparser==2.19
PyMySQL==0.9.3
pyrsistent==0.14.9
pytz==2019.1
PyYAML==5.1
regex==2019.2.3
requests==2.11.1
requests-toolbelt==0.7.0
six==1.12.0
soupsieve==1.9.1
tqdm==4.11.2
tzlocal==1.5.1